+++ README.md
@@ -1,5 +1,2 @@
 # AI_Interviewer_Innov8_3.0_Finals
 This repository contains the codebase for the final problem statement of Innov8_3.0 Hackathon - IIT Delhi 2025
-
-Execution instructions: 
-pip install websockets
